+++ src/api/block_builder/input_selection/manual.rs
@@ -26,7 +26,8 @@ impl<'a> ClientBlockBuilder<'a> {
     /// them with the provided input_range so we can later sign them.
     /// Forwards to [try_select_inputs()] with `force_use_all_inputs` set to true, so all inputs will be included in the
     /// transaction, even if not required for the provided outputs.
-    /// Careful with setting `allow_burning` to `true`, native tokens can get easily burned by accident.
+    /// Careful with setting `allow_burning` to `true`, native tokens, nfts or alias outputs can get easily burned by
+    /// accident.
     pub(crate) async fn get_custom_inputs(
         &self,
         governance_transition: Option<HashSet<AliasId>>,
@@ -36,7 +37,7 @@ impl<'a> ClientBlockBuilder<'a> {
         log::debug!("[get_custom_inputs]");
         let mut inputs_data = Vec::new();
 
-        let local_time = self.client.get_time_checked().await?;
+        let current_time = self.client.get_time_checked().await?;
         if let Some(inputs) = &self.inputs {
             for input in inputs {
                 let output_response = self.client.get_output(input.output_id()).await?;
@@ -46,14 +47,14 @@ impl<'a> ClientBlockBuilder<'a> {
                     let (_output_amount, output_address) = ClientBlockBuilder::get_output_amount_and_address(
                         &output,
                         governance_transition.clone(),
-                        local_time,
+                        current_time,
                     )?;
 
                     let bech32_hrp = self.client.get_bech32_hrp().await?;
-                    let (address_index, internal) = match self.secret_manager {
+                    let address_index_internal = match self.secret_manager {
                         Some(secret_manager) => {
                             match output_address {
-                                Address::Ed25519(_) => {
+                                Address::Ed25519(_) => Some(
                                     search_address(
                                         secret_manager,
                                         &bech32_hrp,
@@ -62,24 +63,28 @@ impl<'a> ClientBlockBuilder<'a> {
                                         self.input_range.clone(),
                                         &output_address,
                                     )
-                                    .await?
-                                }
+                                    .await?,
+                                ),
                                 // Alias and NFT addresses can't be generated from a private key
-                                _ => (0, false),
+                                _ => None,
                             }
                         }
-                        None => (0, false),
+                        // Assuming default for offline signing
+                        None => Some((0, false)),
                     };
+
                     inputs_data.push(InputSigningData {
                         output,
                         output_metadata: OutputMetadata::try_from(&output_response.metadata)?,
-                        chain: Some(Chain::from_u32_hardened(vec![
-                            HD_WALLET_TYPE,
-                            self.coin_type,
-                            self.account_index,
-                            internal as u32,
-                            address_index,
-                        ])),
+                        chain: address_index_internal.map(|(address_index, internal)| {
+                            Chain::from_u32_hardened(vec![
+                                HD_WALLET_TYPE,
+                                self.coin_type,
+                                self.account_index,
+                                internal as u32,
+                                address_index,
+                            ])
+                        }),
                         bech32_address: output_address.to_bech32(&bech32_hrp),
                     });
                 }
@@ -92,7 +97,7 @@ impl<'a> ClientBlockBuilder<'a> {
             self.custom_remainder_address,
             rent_structure,
             allow_burning,
-            local_time,
+            current_time,
         )?;
         Ok(selected_transaction_data)
     }